Remove references to his by renaming to analysis
