+++ src/components/analysis/AnalysisCollection.vue
