+++ src/components/analysis/AnalysisCollectionChart.vue
