+++ src/components/analysis/AnalysisMenu.vue
